requirements.txt
# requirements.txt

amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
kombu==5.5.3
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
